requirements.txt
streamlit>=1.32.0
streamlit-autorefresh>=1.0.1
python-dotenv>=1.0.0
yfinance>=0.2.36
pandas>=2.2.2
numpy>=1.26.4
xgboost>=2.0.3
scikit-learn>=1.4.2
shap>=0.45.1
matplotlib>=3.8.4
seaborn>=0.13.2
nltk>=3.8.1
transformers>=4.40.1
torch>=2.2.2
certifi>=2024.2.2
requests>=2.32.2
